% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assemblage.r
\docType{data}
\name{assemblage}
\alias{assemblage}
\title{Dataset: distribution of 7 archaeological objects across 9 assemblages}
\format{A data frame with 9 rows and 7 variables}
\usage{
data(assemblage)
}
\description{
A dataset containing the frequencies (counts) of 7 objects across 9 assemblages.
}
\keyword{datasets}
