% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deaths.r
\docType{data}
\name{deaths}
\alias{deaths}
\title{Dataset: location of cholera deaths in London (after Dr Snow's mid-1800s study of cholera outbreak in Soho).}
\format{SpatialPointsDataFrame}
\usage{
data(deaths)
}
\description{
A SpatialPointsDataFrame representing the location of cholera deaths in London (after Dr Snow's mid-1800s study of cholera outbreak in Soho).
}
\keyword{datasets}
