% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/destin_loc.r
\docType{data}
\name{destin_loc}
\alias{destin_loc}
\title{Dataset: locations on the volcano Maunga Whau (Auckland, New Zealand)}
\format{SpatialPointsDataFrame}
\usage{
data(destin_loc)
}
\description{
A SpatialPointsDataFrame representing spots on the volcano Maunga Whau (Auckland, New Zealand), to be used as destination locations for least-cost paths calculation.
}
\keyword{datasets}
