% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/etna.r
\docType{data}
\name{etna}
\alias{etna}
\title{Dataset: Digital Terrain Model representing part of the Etna volcano (Sicily, Italy).}
\format{RasterLayer}
\usage{
data(etna)
}
\description{
A RasterLayer representing a Digital Terrain Model (subset of a larger ASTER GDEM representing Sicily) of the Etna volcano (Sicily, Italy).\cr
}
\seealso{
\code{\link{etna_start}} , \code{\link{etna_stop}} , \code{\link{moveCost}}
}
\keyword{datasets}
