% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/etna_start.r
\docType{data}
\name{etna_start}
\alias{etna_start}
\title{Dataset: location on the Etna volcano (Sicily, Italy)}
\format{SpatialPointsDataFrame}
\usage{
data(etna_start)
}
\description{
A SpatialPointsDataFrame representing a spot on the Etna volcano (Sicily, Italy).
}
\seealso{
\code{\link{etna}} , \code{\link{etna_stop}} , \code{\link{moveCost}}
}
\keyword{datasets}
