% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/etna_stop.r
\docType{data}
\name{etna_stop}
\alias{etna_stop}
\title{Dataset: location on the Etna volcano (Sicily, Italy)}
\format{SpatialPointsDataFrame}
\usage{
data(etna_stop)
}
\description{
A SpatialPointsDataFrame representing another spot on the Etna volcano (Sicily, Italy).
}
\seealso{
\code{\link{etna}} , \code{\link{etna_start}} , \code{\link{moveCost}}
}
\keyword{datasets}
