% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/events.r
\docType{data}
\name{events}
\alias{events}
\title{Dataset: location of fictional events}
\format{SpatialPointsDataFrame}
\usage{
data(events)
}
\description{
A SpatialPointsDataFrame representing fictional events.
}
\keyword{datasets}
