% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/faults.r
\docType{data}
\name{faults}
\alias{faults}
\title{Dataset: geological fault-lines in Malta}
\format{SpatialLinesDataFrame}
\usage{
data(faults)
}
\description{
A SpatialLinesDataFrame representing the geological fault-lines in Malta.
}
\keyword{datasets}
