% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impRst.R
\name{impRst}
\alias{impRst}
\title{R function to easily import a raster dataset into R}
\usage{
impRst()
}
\arguments{
\item{none:}{no parameter is used within the function.}
}
\description{
The function is a wrapper for the 'raster()' function out of the 'raster' package. It provides the facility to import
a raster dataset ('RasterLayer' class) by means of a window that allows the user to navigate through the computer's folders and to select the appropriate file.
}
\examples{
my.raster <- impRst() #a window will pop up allowing the user to select the raster dataset

}
\keyword{raster}
