% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locations.r
\docType{data}
\name{locations}
\alias{locations}
\title{Dataset: location of fictional locations}
\format{SpatialPointsDataFrame}
\usage{
data(locations)
}
\description{
A SpatialPointsDataFrame representing fictional locations.
}
\keyword{datasets}
