% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_regr_data.r
\docType{data}
\name{log_regr_data}
\alias{log_regr_data}
\title{Dataset: admission to graduate school}
\format{A data frame with 400 rows and 4 variables}
\usage{
data(log_regr_data)
}
\description{
A dataset containing information about the admission to graduate school of 400 individuals; the dataset features a binary dependent variable and 3 predictors.
Dataset is after: https://stats.idre.ucla.edu/r/dae/logit-regression/
}
\details{
\itemize{
  \item admit. Binary dependent variable (admission yes=1, admission no=0)
  \item gre. Graduate Record Exam score; predictor (continuous)
  \item gpa. Grade Point Average; predictor (continuous)
  \item rank. Prestige of the undergraduate school (ordinal)
}
}
\keyword{datasets}
