% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/malta_polyg.r
\docType{data}
\name{malta_polyg}
\alias{malta_polyg}
\title{Dataset: Malta polygon}
\format{SpatialPolygonsDataFrame}
\usage{
data(malta_polyg)
}
\description{
A SpatialPolygonsDataFrame representing Malta.\cr
}
\keyword{datasets}
