% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Massachusetts.r
\docType{data}
\name{Massachusetts}
\alias{Massachusetts}
\title{Dataset: Massachusetts state limit}
\format{SpatialPolygonsDataFrame}
\usage{
data(Massachusetts)
}
\description{
A SpatialPolygonsDataFrame representing the limits of Massachusetts.\cr
After \url{https://mgimond.github.io/Spatial/point-pattern-analysis-in-r.html}.
}
\keyword{datasets}
