% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phases.r
\docType{data}
\name{phases}
\alias{phases}
\title{Dataset: Posterior Probabilities for the chronological relation of the Starting and Ending boundaries of two Bayesian independent 14C phases}
\format{A data frame with 4 rows and 5 variables}
\usage{
data(phases)
}
\description{
A dataset containing the posterior probability (as calculated by the OxCal program) for the relative chronological relations for the
starting and ending boundaries of two Bayesian-defined 14C phases.
}
\keyword{datasets}
