% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/points.r
\docType{data}
\name{points}
\alias{points}
\title{Dataset: location of fictional points}
\format{SpatialPointsDataFrame}
\usage{
data(points)
}
\description{
A SpatialPointsDataFrame representing fictional locations.
}
\keyword{datasets}
