% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pointsInPolygons.R
\name{pointsInPolygons}
\alias{pointsInPolygons}
\title{R function to test points-in-polygons association}
\usage{
pointsInPolygons(point.feat, polyg.feat, buffer = 0)
}
\arguments{
\item{point.feat:}{feature (of point type) whose spatial association with the polygons has to be assessed.}

\item{polyg.feat:}{feature (polygon type) in relation to which the spatial association of the points has to be assessed.}

\item{buffer:}{add a buffer to the convex hull of the study area (0 by default); the unit depends upon the units of the input data.}
}
\description{
The function allows to test if there is a significant spatial association between a set of points and a set of polygons, in terms of points falling within the polygons.
In other words, it aims at testing whether a set of points falls inside a set of polygons more often than would be expected by chance.
The basic assumption is that the polygons are completely contained within the study plot.
The calculations are based on the bounding polygon based on the union the convex hulls of the point and of the polygon feature. The bounding polygon is considered as representing the study plot itself.
}
\details{
The computational bases of the function are described in the help documentation of the 'Point-Polygon Relationship' analysis facility provided by the PASSaGE software (http://www.passagesoftware.net/manual.php).\cr

The function makes use of the 'dbinom() and 'pbinom()' functions.\cr
The probability of observed count within polygons is 'dbinom(x, size=n.of.points, prob=p)', where 'x' is the observed number of points within polygons, 'n.of.points' is the total number of points, and 'p' is the probability that a single point will be found within a polygon, which is equal to the ratio between the area of the polygons and the total area of the study plot.\cr
The probability that x or fewer points will be found within the polygons is 'pbinom(x, size=n.of.points, prob=p)'.

The function produces a plot of the points and polygons (plus the convex hull of the study area), and relevant information are reported at the bottom of the chart itself.\cr

A list is also returned, containing what follows:\cr
-$Polygons' area;\cr
-$Study area's area;\cr
-$Total # of points;\cr
-$Observed # of points in polygons;\cr
-$Expected # of points in polygons;\cr
-$Exact probability of observed count within polygons;\cr
-$Probability of <= observed count within polygons;\cr
-$Probability of >= observed count within polygons.
}
\examples{
data(points)
data(polygons)
result <- pointsInPolygons(points, polygons)

}
\keyword{association}
