% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polygons.r
\docType{data}
\name{polygons}
\alias{polygons}
\title{Dataset: location of fictional polygons}
\format{SpatialPolygonsDataFrame}
\usage{
data(polygons)
}
\description{
A SpatialPolygonsDataFrame representing fictional polygons.
}
\keyword{datasets}
