% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/popdensity.r
\docType{data}
\name{popdensity}
\alias{popdensity}
\title{Dataset: Massachusetts population density}
\format{RasterLayer}
\usage{
data(popdensity)
}
\description{
A RasterLayer representing the population density in Massachusetts.\cr
After \url{https://mgimond.github.io/Spatial/point-pattern-analysis-in-r.html}.
}
\keyword{datasets}
