% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppd_plot.r
\name{ppdPlot}
\alias{ppdPlot}
\title{R function for plotting Posterior Probability Densities for Bayesian modeled 14C dates/parameters}
\usage{
ppdPlot(data, lower = min(data[, 1]), upper = max(data[, 1]), type)
}
\arguments{
\item{data:}{dataframe containing the data as returned by the OxCal program.}

\item{lower:}{lower limit of the calendar date axis.}

\item{upper:}{upper limit of the calendar date axis; if the lower and upper parameters are not provided, the default values will be the earliest and latest calendar dates.}

\item{type:}{type of plot the user wishes to plot (a: curves outlined by a line; b: curves plotted as solid areas; c: combination of a and b).}
}
\description{
The function allows plot Posterior Probability Densities with a nice outlook thanks to 'ggplot2'.\cr
It takes as input a dataframe that must be organized as follows (it is rather easy to do that once the data have been exported from OxCal):\cr
-calendar dates (first column to the left);\cr
-posterior probabilities (second column);\cr
-grouping variables (third column), which could contain the names of the events of interest (e.g., phase 1 start, phase 1 end, phase 2 start, phase 2 end, etc).
}
\examples{
data(radioc_data) #load a toy dataset
ppdPlot(radioc_data, type="a") #plot the Posterior Probability Densities for the phases' parameters
ppdPlot(radioc_data, -1000, 100, type="b") #plot the Posterior Probability Densities for the phases' parameters, setting different boundaries for the x-axis and using filled curves instead of simple outlines

}
\keyword{Bayesian}
\keyword{dates}
\keyword{density}
\keyword{parameters}
\keyword{posterior}
\keyword{probability}
\keyword{radiocarbon}
