% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prob_phases_relat.R
\name{prob.phases.relat}
\alias{prob.phases.relat}
\title{R function to calculate the Posterior Probability for different chronological relations between two Bayesian radiocarbon phases}
\usage{
prob.phases.relat(data, sort = FALSE)
}
\arguments{
\item{data:}{matrix containing the posterior probability of the chronological relation between the Starting and Ending boundaries of two independent phases, as returned by the OxCal's 'Order' query (see Deatils).}

\item{sort:}{logical which takes TRUE or FALSE (default) if the user does or does not want the returned posterior probabilities sorted in descending order.}
}
\description{
The function allows to calculate the posterior probability for different chronological relations between two phases defined via Bayesian radiocarbon modeling.
For the results to make sense, the phases have to be defined as independent if one wishes to assess what is the posterior probability for different relative chronological relations between them.\cr
}
\details{
The rationale for this approach is made clear in an article by Buck et al 1992 (https://doi.org/10.1016/0305-4403(92)90025-X), and it runs as follows:
"if we do not make any assumption about the relationship between the phases, can we test how likely they are to be in any given order"?\cr

The function takes as input the table provided by the 'OxCal' program  as result of the 'Order' query.
Once the table as been saved from 'OxCal' in .csv format, you have to feed it in R. A .csv file can be imported into R using (for instance):\cr
\eqn{mydata <- read.table(file.choose(), header=TRUE, sep=",", dec=".", as.is=T)}\cr

Make sure to insert the phases' parameters (i.e., the starting and ending boundaries of the two phases) in the OxCal's Order query in the following order:
StartA, EndA, StartB, EndB. That is, first the start and end of your first phase, then the start and end of the second one. You can give any name to your phases, as long as the order is like the one described.\cr

Given two phases A and B, the function allows to calculate the posterior probability for:\cr
-A being within B\cr
-B being within A\cr
-A starting earlier but overlapping with B\cr
-B starting earlier but overlapping with A\cr
-A being entirely before B\cr
-B being entirely before A\cr
-sA being within B\cr
-eA being within B\cr
-sB being within A\cr
-eB being within A\cr
where 's' and 'e' refer to the starting and ending boundaries of a phase. The function will return a table and a dot plot.\cr
Thanks are due to Dr. Andrew Millard (Durham University) for the help provided in working out the operations on probabilities.
}
\examples{
data(phases) #load a toy dataset
res <- prob.phases.relat(phases) #calculate the Posterior Probability for the chronological relation between two phases, stores the results in the 'res' object, and produce a dot chart.

}
\keyword{radiocarbon}
