% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pumps.r
\docType{data}
\name{pumps}
\alias{pumps}
\title{Dataset: location of public water pumps in London (after Dr Snow's mid-1800s study of cholera outbreak in Soho).}
\format{SpatialPointsDataFrame}
\usage{
data(pumps)
}
\description{
A SpatialPointsDataFrame representing SpatialPointsDataFrame representing the location of public water pumps in London (after Dr Snow's mid-1800s study of cholera outbreak in Soho).
}
\keyword{datasets}
