% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radioc_data.r
\docType{data}
\name{radioc_data}
\alias{radioc_data}
\title{Dataset: Posterior Probabilities for the Starting and Ending boundaries of two 14C phases}
\format{A data frame with 680 rows and 3 variables}
\usage{
data(radioc_data)
}
\description{
A dataset containing the calendar dates and the associated posterior probability for the starting and ending boundaries of two 14C phases.
}
\details{
\itemize{
  \item dates. Calendar dates
  \item prob. Posterior probability
  \item group. Grouping variable containing strings referring to the start and ending boundaries of the two 14C phases
}
}
\keyword{datasets}
