% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rndpoints.r
\docType{data}
\name{rndpoints}
\alias{rndpoints}
\title{Dataset: random points}
\format{SpatialPointsDataFrame}
\usage{
data(rndpoints)
}
\description{
A SpatialPointsDataFrame representing random locations.
}
\keyword{datasets}
