% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Starbucks.r
\docType{data}
\name{Starbucks}
\alias{Starbucks}
\title{Dataset: location of Starbucks in Massachusetts}
\format{SpatialPointsDataFrame}
\usage{
data(Starbucks)
}
\description{
A SpatialPointsDataFrame representing the location of Starbucks in Massachusetts.\cr
After \url{https://mgimond.github.io/Spatial/point-pattern-analysis-in-r.html}.
}
\keyword{datasets}
