% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thiessenpolyg.r
\docType{data}
\name{thiessenpolyg}
\alias{thiessenpolyg}
\title{Dataset: Thiessen polygons around the points represented in the 'locations' dataset}
\format{SpatialPolygonsDataFrame}
\usage{
data(thiessenpolyg)
}
\description{
A SpatialPolygonsDataFrame representing Thiessen polygons around the points represented in the 'locations' dataset.
}
\keyword{datasets}
