% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volc.r
\docType{data}
\name{volc}
\alias{volc}
\title{Dataset: Digital Terrain Model representing the volcano Maunga Whau (Auckland, New Zealand)}
\format{RasterLayer}
\usage{
data(volc)
}
\description{
A RasterLayer representing a Digital Terrain Model of the volcano Maunga Whau (Auckland, New Zealand).\cr
}
\keyword{datasets}
