% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brsim.r
\name{BRsim}
\alias{BRsim}
\title{R function for Brainerd-Robinson simiarity coefficient}
\usage{
BRsim(data, which = "rows", correction = FALSE, rescale = TRUE)
}
\arguments{
\item{data:}{dataframe containing the dataset (note: assemblages in rows, variables in columns).}

\item{which:}{takes "rows" (default) if the user wants the coefficients be calculated for the row categories, "cols" if the users wants the coefficients be calculated for the column categories.}

\item{correction:}{takes FALSE (default) if the user does not want the coefficients to be corrected, while TRUE will provide corrected coefficients.}

\item{rescale:}{takes FALSE if the user does NOT want the coefficients to be rescaled between 0.0 and 1.0 (i.e., the user will get the original version of the Brainerd-Robinson coefficient (spanning from 0 [maximum dissimilarity] to 200 [maximum similarity]), while TRUE (default) will return rescaled coefficient.}
}
\description{
The function allows to calculate the Brainerd-Robinson similarity coefficient, taking as input a cross-tabulation (dataframe).\cr
}
\details{
The function returns:\cr
a) a correlation matrix in tabular form;\cr
b) a heat-map representing, in a graphical form, the aforementioned correlation matrix.\cr

In the heat-map (which is built using the 'corrplot' package), the size and the color of the squares are proportional to the Brainerd-Robinson coefficients, which are also reported by numbers.\cr

In order to "penalize" BR similarity coefficient(s) arising from assemblages with unshared categories, the function does what follows: it divides the BR coefficient(s) by the number of unshared categories plus 0.5. The latter addition is simply a means to be still able to penalize coefficient(s) arising from assemblages having just one unshared category. Also note that joint absences will have no weight on the penalization of the coefficient(s). In case of assemblages sharing all their categories, the corrected coefficient(s) turns out to be equal to the uncorrected one.
}
\examples{
data(assemblage)
coeff <- BRsim(data=assemblage, correction=FALSE, rescale=TRUE)

}
\keyword{AUC}
\keyword{logistic}
\keyword{model}
\keyword{optimism}
\keyword{regression}
\keyword{validation}
