% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impShp.R
\name{impShp}
\alias{impShp}
\title{R function to easily import a vectorial dataset (shapefile) into R}
\usage{
impShp()
}
\arguments{
\item{none:}{no parameter is used within the function.}
}
\description{
The function is a wrapper for the 'shapefile()' function out of the 'raster' package. It provides the facility to import
a vectorial dataset (of shapefile type) by means of a window that allows the user to navigate through the computer's folders and to select the appropriate file.
}
\examples{
my.shapefile <- impShp() #a window will pop up allowing the user to select the shapefile

}
\keyword{shapefile}
