% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_valid.r
\name{modelvalid}
\alias{modelvalid}
\title{R function for binary Logistic Regression internal validation}
\usage{
modelvalid(data, fit, B)
}
\arguments{
\item{data:}{dataframe containing the dataset (Dependent Variable must be stored in the first column to the left).}

\item{fit:}{object returned from glm() function.}

\item{B:}{desired number of iterations (see description of the procedure above).}
}
\description{
The function allows to perform internal validation of a binary Logistic Regression model, implementing part of the procedure described in
Arboretti Giancristofaro R, Salmaso L. "Model performance analysis and model validation in logistic regression". Statistica 2003(63): 375–396.\cr
The procedure consists of the following steps:\cr
(1) the whole dataset is split into two random parts, a fitting (75 percent) and a validation (25 percent) portion;\cr
(2) the model is fitted on the fitting portion (i.e., its coefficients are computed considering only the observations in that portion) and its performance is evaluated on both the fitting and the validation portion, using AUC as performance measure;\cr
(3) steps 1-2 are repeated B times, eventually getting a fitting and validation distribution of the AUC values. The former provides an estimate of the performance of the model in the population of all the theoretical training samples; the latter represents an estimate of the model’s performance on new and independent data.\cr
The function returns two boxplots that represent the training and the testing (i.e., validation) distribution of the AUC value across the 1000 iterations. For an example of the interpretation of the chart, see the aforementioned article, especially page 390-91.
}
\examples{
data(log_regr_data) # load the sample dataset
model <- glm(admit ~ gre + gpa + rank, data = log_regr_data, family = "binomial") # fit a logistic regression model, storing the results into an object called 'model'
modelvalid(data=log_regr_data, fit=model, B=1000) # run the function, using 1000 iterations

}
\keyword{bootstrap}
\keyword{logistic}
\keyword{regression}
\keyword{validation}
