% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monthlyWind.R
\name{monthlyWind}
\alias{monthlyWind}
\title{R function to download month-averaged wind data (from Global Forecast System (GFS) of the USA's National Weather Service (NWS))}
\usage{
monthlyWind(raster, year = 2015, month = 1, days = 31, lon1 = NULL,
  lon2 = NULL, lat1 = NULL, lat2 = NULL)
}
\arguments{
\item{raster:}{raster dataset representing the study area.}

\item{year:}{selected year (from 2011 to current); 2015 by default.}

\item{month:}{selected month; 01 by default.}

\item{days:}{number of days featuring the selected month; 31 by default.}

\item{lon1:}{western longitude. If it is set to NULL (default), it will be derived from the input raster.}

\item{lon2:}{eastern longitude. If it is set to NULL (default), it will be derived from the input raster.}

\item{lat1:}{southern latitude. If it is set to NULL (default), it will be derived from the input raster.}

\item{lat2:}{northern latitude. If it is set to NULL (default), it will be derived from the input raster.}
}
\description{
The function allows to download wind data from NOAA/NCEP Global Forecast System (GFS) Atmospheric Model colection, creating monthly averages.
The extent of the study area can be specified either by entering the geographic coordinates, or on the basis of an input raster dataset representing the study area itself (for instance, a Digital Elevation Model).
The function saves two .geotiff files in the computer's working directory, one representing the wind speed, the other the wind direction. In both cases, the values are the average of the wind speed and direction values in the study area across the days of the selected month, in the selected year. A plot is also returned in the R console.\cr
The function returns a list containing the following data:\cr
-$windMonth: stores the U and V components for each output grid cell (spatial resolution 0.5 degrees=50 Km)\cr
-$windMonthFit: stores the wind speed and direction for each output grid cell.\cr
The function builds upon the 'wind.dl()' function from Javier Fern치ndez-L칩pez's package 'rWind'. The help provided by Dr Fern치ndez-L칩pez in creating an earlier version of the 'monthlyWind()' function is gratefully acknowledged.
}
\examples{
res <- monthlyWind(year=2014, month=12, days=31, lon1=-10, lon2=5, lat1=35, lat2=45) #download wind data for Spain region, averaging the values across the 31 days of December 2014

}
\keyword{wind}
