% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refNNa.R
\name{refNNa}
\alias{refNNa}
\title{R function for refined Nearest Neighbor analysis of point patterns (G function)}
\usage{
refNNa(feature, studyplot = NULL, buffer = 0, B = 1000, order = 1)
}
\arguments{
\item{feature:}{feature dataset (of point type).}

\item{studyplot:}{shapefile (of polygon type) representing the study area; if not provided, the study area is internally worked out as the convex hull enclosing the input feature dataset.}

\item{buffer:}{add a buffer to the studyplot (0 by default); the unit depends upon the units of the input data.}

\item{B:}{number of randomizations to be used (1000 by default).}

\item{order:}{integer indicating the kth nearest neighbour (1 by default).}
}
\description{
The function allows to perform the refined Nearest Neighbor analysis of point patterns by plotting the cumulative Nearest Neighbour distance,
along with a 95percent confidence envelope and a curve representing the expected cumulative distribution under the assumption of complete spatial randomness.\cr
}
\details{
The function uses a randomized approach to build the confidence envelope, whereby cumulative distributions of average NN distances of random points are computed across B iterations (1000 by default).
In each iteration, a set of random points (with sample size equal to the number of points of the input feature) is drawn.\cr
}
\examples{
data(springs)
refNNa(springs) #produces a plot representing the cumulative nearest neighbour distance distribution; a confidence envelope based on 1000 ranodmized simulations is also shown.
}
\seealso{
\code{\link{NNa}}
}
\keyword{refNNA}
