% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robust_BA_plot.R
\name{robustBAplot}
\alias{robustBAplot}
\title{R function to plot a robust version of the Bland-Altman plot}
\usage{
robustBAplot(a, b, z = 1.96, meth_A_lab = "Method A",
  meth_B_lab = "Method B", cex = 0.6)
}
\arguments{
\item{a:}{vector storing the first set of measurements to be compared.}

\item{b:}{vector storing the second set of measurements to be compared.}

\item{z:}{value for the confidence interval for the median difference; set by default to 1.96 (corresponding to 95 percent CI).}

\item{meth_A_lab:}{label to be used in the returned chart to refer to the method yielding the first set of measurements.}

\item{meth_B_lab:}{label to be used in the returned chart to refer to the method yielding the second set of measurements.}

\item{cex:}{size of the data points.}
}
\description{
The function allows to plot a robust version of the Bland-Altman plot.\cr
The chart is based on robust (i.e. resistant to outlying values) measures of central tendency and variability: median and Median Absolute Deviation (MAD) (Wilcox R R. 2001. "Fundamentals of modern statistical methods: Substantially improving power and accuracy". New York: Springer) instead of mean and standard deviation.\cr
The x-axis displays the median of the estimated and reported duration, while the y-axis represents their difference. A solid horizontal line represents the bias, i.e. the median of the differences reported on the y-axis. Two dashed horizontal lines represent the region in which 95percent of the observations are expected to lie; they are set at the median plus or minus z*(MAD/0.6745).
}
\examples{
a <- rnorm(30,10,1) #create a first toy vector
b <- a*runif(30,1,1.5) #create a second toy vector
robustBAplot(a,b) #returns the robust BA plot

}
\keyword{Altman}
\keyword{Bland}
\keyword{MAD}
\keyword{median}
\keyword{plot}
\keyword{robust}
